+++ packages/logger/src/winston-logger/adapter.ts
@@ -2,7 +2,14 @@ import type { Transport, TransportOptions } from '../types';
 import { LogLevel } from '../types';
 import type WinstonTransport from 'winston-transport';
 import winston, { format } from 'winston';
-import { ConsoleTransport, FileTransport, NullTransport, UI5ToolingTransport, VSCodeTransport } from '../transports';
+import {
+    ConsoleTransport,
+    ArrayTransport,
+    FileTransport,
+    NullTransport,
+    UI5ToolingTransport,
+    VSCodeTransport
+} from '../transports';
 import { NullTransport as WinstonNullTransport } from './null-transport';
 import { VSCodeTransport as WinstonVSCodeTransport } from './vscode-output-channel-transport';
 import type { Format } from 'logform';
@@ -96,7 +103,7 @@ const consoleFormat = format.combine(
  *  Will throw an error if the transport is not recognized
  *
  * @param transport
- * @returns {WinstonTransport}
+ * @returns {WinstonTransport} winston transport
  */
 export function toWinstonTransport(transport: Transport): WinstonTransport {
     if (transport instanceof NullTransport) {
@@ -115,6 +122,8 @@ export function toWinstonTransport(transport: Transport): WinstonTransport {
             ...toWinstonTransportOptions(transport.options),
             format: ui5ToolingFormat(transport.options.moduleName)
         });
+    } else if (transport instanceof ArrayTransport) {
+        return transport;
     } else {
         throw new Error('Unrecognized transport type');
     }