+++ packages/btp-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/btp-utils
 
+## 0.11.0
+
+### Minor Changes
+
+-   bc4cb3a: New module @sap-ux/environment-check. Adds additional destination property 'HTML5.DynamicDestination' in @sap-ux/btp-utils
+
 ## 0.10.4
 
 ### Patch Changes