+++ packages/ui5-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/ui5-proxy-middleware
 
+## 1.1.8
+
+### Patch Changes
+
+-   Updated dependencies [bc4cb3a]
+    -   @sap-ux/logger@0.3.0
+
+## 1.1.7
+
+### Patch Changes
+
+-   Updated dependencies [b6d0c67]
+    -   @sap-ux/ui5-config@0.14.5
+
 ## 1.1.6
 
 ### Patch Changes